+++ build-logic/settings.gradle.kts
@@ -1,0 +1,9 @@
+dependencyResolutionManagement {
+  repositories {
+    gradlePluginPortal()
+    mavenCentral()
+  }
+  versionCatalogs { create("libs") { from(files("../gradle/libs.versions.toml")) } }
+}
+
+rootProject.name = "build-logic"
