+++ build-logic/src/main/kotlin/dbos.publishing-conventions.gradle.kts
@@ -1,0 +1,46 @@
+import com.vanniktech.maven.publish.DeploymentValidation
+
+plugins { id("com.vanniktech.maven.publish") }
+
+val publishingToMavenCentral =
+  gradle.startParameter.taskNames.any { it.contains("publishToMavenCentral") }
+
+tasks.withType<Javadoc> {
+  (options as StandardJavadocDocletOptions).apply {
+    addStringOption("Xdoclint:all,-missing", "-quiet")
+    encoding = "UTF-8"
+  }
+}
+
+tasks.named("build") { dependsOn("javadoc") }
+
+extensions.configure<com.vanniktech.maven.publish.MavenPublishBaseExtension> {
+  publishToMavenCentral(automaticRelease = true, validateDeployment = DeploymentValidation.NONE)
+  if (publishingToMavenCentral) signAllPublications()
+
+  pom {
+    inceptionYear.set("2025")
+    url.set("https://github.com/dbos-inc/dbos-transact-java")
+
+    licenses {
+      license {
+        name.set("MIT License")
+        url.set("https://opensource.org/licenses/MIT")
+      }
+    }
+
+    developers {
+      developer {
+        id.set("dbos-inc")
+        name.set("DBOS Inc")
+        email.set("support@dbos.dev")
+      }
+    }
+
+    scm {
+      connection.set("scm:git:git://github.com/dbos-inc/dbos-transact-java.git")
+      developerConnection.set("scm:git:ssh://github.com:dbos-inc/dbos-transact-java.git")
+      url.set("https://github.com/dbos-inc/dbos-transact-java/tree/main")
+    }
+  }
+}
