+++ gradle/libs.versions.toml
@@ -15,6 +15,7 @@ junit = "6.1.0"
 junit-pioneer = "2.3.0"
 kotlin = "2.3.21"
 kryo = "5.6.2"
+ktfmt = "0.62"
 logback = "1.5.33"
 maven-artifact = "3.9.16"
 maven-publish = "0.36.0"
